Separating the UI from the Game Controller
