+++ Assets/Prefabs/Player.prefab
@@ -119,7 +119,6 @@ GameObject:
   - 61: {fileID: 61000011862802682}
   - 50: {fileID: 50000011862695246}
   - 58: {fileID: 58000012316540882}
-  - 212: {fileID: 212000013056764892}
   - 114: {fileID: 114000013579520362}
   - 114: {fileID: 114000013599799802}
   - 95: {fileID: 95000011927050182}
@@ -223,7 +222,7 @@ Transform:
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1000013470291786}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0.12, y: 4.3, z: -9}
+  m_LocalPosition: {x: -0.0288634, y: 4.5293846, z: -9}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_Children:
@@ -406,9 +405,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 19d7614d8a623224ca772a2308a591ee, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  resistance: 1
   xAccel: 4
   maxSpeed: 20
   jumpForce: 3500
+  startingHealth: 100
+  currentHealth: 0
   currentWeapon: {fileID: 0}
   defaultWeapon: {fileID: 114000013048013508, guid: 1838a38f20b2f074fa1304dd8c7caa41,
     type: 2}
@@ -417,14 +419,6 @@ MonoBehaviour:
   whatIsGround:
     serializedVersion: 2
     m_Bits: 4352
-  startingHealth: 100
-  currentHealth: 0
-  hasPowerUp: 0
-  powerUpTimer: 0
-  powerUpName: 
-  speedMultiplier: 0
-  jumpMultiplier: 0
-  powerUpWeapon: {fileID: 0}
   blink_speed: 0.3
   invincible_duration: 3
 --- !u!114 &114000013599799802
@@ -547,39 +541,6 @@ SpriteRenderer:
   m_SortingLayerID: 0
   m_SortingOrder: 0
   m_Sprite: {fileID: 21300010, guid: e6ef5b64103b513449e7559c211fe534, type: 3}
-  m_Color: {r: 1, g: 1, b: 1, a: 1}
-  m_FlipX: 0
-  m_FlipY: 0
---- !u!212 &212000013056764892
-SpriteRenderer:
-  m_ObjectHideFlags: 1
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1000013470291786}
-  m_Enabled: 0
-  m_CastShadows: 0
-  m_ReceiveShadows: 0
-  m_MotionVectors: 1
-  m_LightProbeUsage: 0
-  m_ReflectionProbeUsage: 0
-  m_Materials:
-  - {fileID: 10754, guid: 0000000000000000f000000000000000, type: 0}
-  m_SubsetIndices: 
-  m_StaticBatchRoot: {fileID: 0}
-  m_ProbeAnchor: {fileID: 0}
-  m_LightProbeVolumeOverride: {fileID: 0}
-  m_ScaleInLightmap: 1
-  m_PreserveUVs: 0
-  m_IgnoreNormalsForChartDetection: 0
-  m_ImportantGI: 0
-  m_SelectedWireframeHidden: 1
-  m_MinimumChartSize: 4
-  m_AutoUVMaxDistance: 0.5
-  m_AutoUVMaxAngle: 89
-  m_LightmapParameters: {fileID: 0}
-  m_SortingLayerID: 0
-  m_SortingOrder: 0
-  m_Sprite: {fileID: 21300000, guid: 925a4c2bc43d4d846be373ebb55d28f0, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
