+++ Assets/Scripts/GameController.cs
@@ -1,22 +1,10 @@
 ﻿using UnityEngine;
 using System.Collections;
-using UnityEngine.UI;
 using UnityEngine.SceneManagement;
 
 public class GameController : MonoBehaviour {
 
-	public Text scoreText;
-	public Text timerText;
-	public Text roundText;
-	public Text ammoText;
-	public Text countdownText;
 	private float countdownTimer;
-	public Text powerUpText;
-	/* 
-	 * This is the rect transform for the green health because that is what
-	 * 	we are modifying
-	 */
-	public RectTransform healthBar;
 
 	// To determine which controller is which player
 	// This is for the purpose of swapping roles
@@ -26,7 +14,6 @@ public class GameController : MonoBehaviour {
 	private int score = 0;
 	//Increase this for a longer Creator phase
 	private float timer = 10.0F;
-	private float powerUpTimer;
 	private int state;
 	private int round;
 
@@ -40,17 +27,11 @@ public class GameController : MonoBehaviour {
 	public PlayerController playerPrefab;
 	private CreatorController creator;
 	private PlayerController player;
-	private WeaponController shoot;
 	private DynamicCamera camera;
 
 	void Start () {
-		scoreText.text = "";
-		ammoText.text = "";
-		healthBar.parent.gameObject.SetActive (false);
 		state = 0;
 		round = 1;
-		width = healthBar.rect.width;
-		startMaxXPos = healthBar.offsetMax.x;
 		camera = GameObject.Find("Main Camera").GetComponent<DynamicCamera>();
 		generateMap ();
 	}
@@ -62,27 +43,22 @@ public class GameController : MonoBehaviour {
 				if (!creator) {
 					createCreator ();
 				}
-				updateTimer (true);
 				if (timer <= 0) {
 					DestroyObject (creator.gameObject);
 					timer = phaseSwitchTimes[0];
-					countdownText.text = phaseSwitchMessages [0];
 					state = 1;
 				}
 				break;
 			}
 		case 1: //Phase Switch
 			{
-				updateTimer (false);
 				if (timer <= 0) {
 					phaseSwitchState++;
 					if (phaseSwitchState >= phaseSwitchMessages.Length) {
-						countdownText.text = "";
 						createPlayer ();
 						timer = 120.0F;
 						state = 2;
 					} else {
-						countdownText.text = phaseSwitchMessages [phaseSwitchState];
 						timer = phaseSwitchTimes [phaseSwitchState];
 					}
 				}
@@ -90,10 +66,6 @@ public class GameController : MonoBehaviour {
 			}
 		case 2: //Player
 			{
-				updateTimer (true);
-				// Displays current statistics for player in UI labels:
-				scoreText.text = score.ToString ();
-				ammoText.text = player.currentWeapon.ammo.ToString ();
 				if (timer <= 0 || player.currentHealth <= 0) {
 					state = 3;
 				}
@@ -101,62 +73,19 @@ public class GameController : MonoBehaviour {
 			}
 		case 3: //TODO: End of Round
 			{
-				// Removes everything from UI:
-				scoreText.text = "";
-				timerText.text = "";
-				roundText.text = "";
-				ammoText.text = "";
-				countdownText.text = "Round over!";
 				break;
 			}
 		}
-		roundText.text = "Round: " + round;
 	}
 
 	private void createPlayer() {
 		player = Instantiate (playerPrefab);
 		camera.setFollowing (player.gameObject);
-		healthBar.parent.gameObject.SetActive (true);
 	}
 
 	private void createCreator() {
 		creator = Instantiate (creatorPrefab);
 		camera.setFollowing (creator.gameObject);
-		healthBar.parent.gameObject.SetActive (false);
-	}
-
-	private void updateTimer(bool showText) {
-		timer = timer - Time.deltaTime;
-		if (powerUpTimer > 0) powerUpTimer -= Time.deltaTime;
-		if (showText){
-			timerText.text = (int)((timer + 1) / 60) + ":" + (int)(((timer + 1) % 60) / 10) + (int)(((timer + 1) % 60) % 10);
-			if (powerUpTimer > 0)
-				powerUpText.text =	"" + 
-									(int)(((powerUpTimer + 1) % 60) / 10) + 
-									(int)(((powerUpTimer + 1) % 60) % 10);
-			roundText.text = "Round: " + round;
-		} else {
-			// Removes everything from UI:
-			scoreText.text = "";
-			timerText.text = "";
-			roundText.text = "";
-			ammoText.text = "";
-			powerUpText.text = "";
-			timerText.text = "";
-		}
-
-	}
-
-	public void applyPowerUp(float duration, Sprite image)
-	{
-		GameObject.Find("PowerUpDisplay").GetComponentInChildren<Image>().sprite = image;
-		powerUpTimer = duration;
-	}
-
-	// Update the UI's health to correctly reflect the current player's health
-	public void updateHealth () {
-		float playerHealth = player.currentHealth / player.startingHealth;
-		healthBar.localScale = new Vector2 (Mathf.Clamp(playerHealth, 0F, 1F), healthBar.localScale.y);
 	}
 
 	public void generateMap(){
