+++ Assets/Scripts/Player_Scripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 0954baa95c20ef345a64b161f7e120b7
+folderAsset: yes
+timeCreated: 1478472066
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
