+++ Assets/Scripts/UserInterfaces.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d3d2d75b0cc2d5b4293499565d879cd0
+folderAsset: yes
+timeCreated: 1478472290
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
