+++ Assets/Scripts/UserInterfaces/PlayerHud.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c9de682d9e0453249b3749776a2e3927
+timeCreated: 1478324989
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
